requirements.txt
streamlit
pillow
streamlit-lottie